# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/messages:
    get:
      tags:
        - Messages
      summary: List Messages
      description: Get a paginated list of Messages with optional filtering and sorting
      operationId: listMessages
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Messages
      summary: Create Message
      description: Create a new Message
      operationId: createMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/threads:
    get:
      tags:
        - Threads
      summary: List Threads
      description: Get a paginated list of Threads with optional filtering and sorting
      operationId: listThreads
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Threads
      summary: Create Thread
      description: Create a new Thread
      operationId: createThread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThreadRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thread'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Message:
      type: object
      required:
        - id
        - thread_id
        - direction
        - channel
        - body
        - status
        - occurred_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        thread_id:
          type: string
          format: uuid
        direction:
          type: string
        channel:
          type: string
        external_id:
          nullable: true
          type: string
        address_from:
          nullable: true
          type: string
        address_to:
          nullable: true
          type: string
        body:
          type: string
        status:
          type: string
        failure_reason:
          nullable: true
          type: string
        occurred_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateMessageRequest:
      type: object
      required:
        - thread_id
        - direction
        - channel
        - body
        - status
        - occurred_at
        - metadata
      properties:
        thread_id:
          type: string
          format: uuid
        direction:
          type: string
        channel:
          type: string
        external_id:
          nullable: true
          type: string
        address_from:
          nullable: true
          type: string
        address_to:
          nullable: true
          type: string
        body:
          type: string
        status:
          type: string
        failure_reason:
          nullable: true
          type: string
        occurred_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    MessageList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Thread:
      type: object
      required:
        - id
        - channel
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        channel:
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        subject_type:
          nullable: true
          type: string
        subject_id:
          nullable: true
          type: string
          format: uuid
        external_ref:
          nullable: true
          type: string
        status:
          type: string
        last_message_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateThreadRequest:
      type: object
      required:
        - channel
        - status
        - metadata
      properties:
        channel:
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        subject_type:
          nullable: true
          type: string
        subject_id:
          nullable: true
          type: string
          format: uuid
        external_ref:
          nullable: true
          type: string
        status:
          type: string
        last_message_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    ThreadList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Thread'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []