# src/presentation/http/message_handler.rs
//! Message REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use chrono::{DateTime, Utc};
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::application::service::{MessageService, ServiceError};
use crate::domain::entity::*;

// DTO imports
use crate::presentation::dto::{
    CreateMessageDto, MessageResponseDto, PatchMessageDto, UpdateMessageDto,
};

/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum MessageError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for MessageError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for MessageError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "MESSAGE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "MESSAGE_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "MESSAGE_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "MESSAGE_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for Message.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /messages | List with pagination |
/// | POST | /messages | Create new |
/// | GET | /messages/:id | Get by ID |
/// | PUT | /messages/:id | Full update |
/// | PATCH | /messages/:id | Partial update |
/// | DELETE | /messages/:id | Soft delete |
/// | POST | /messages/bulk | Bulk create |
/// | POST | /messages/upsert | Upsert |
/// | GET | /messages/trash | List deleted |
/// | POST | /messages/:id/restore | Restore |
/// | DELETE | /messages/empty | Empty trash |
/// | GET | /messages/:id/deleted | Get deleted by ID |
/// | DELETE | /messages/trash/:id | Permanent delete from trash |
/// | GET | /messages/count | Count active entities |
/// | GET | /messages/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(MessageService::with_repository(repository));
/// let router = create_message_routes(service);
/// ```
pub fn create_message_routes(service: Arc<MessageService>) -> Router {
    BackboneCrudHandler::<
        MessageService,
        Message,
        CreateMessageDto,
        UpdateMessageDto,
        MessageResponseDto,
    >::routes(service, "/messages")
}

/// Create Axum router with only the read (GET) endpoints for Message.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_message_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_message_read_routes(service: Arc<MessageService>) -> Router {
    BackboneCrudHandler::<
        MessageService,
        Message,
        CreateMessageDto,
        UpdateMessageDto,
        MessageResponseDto,
    >::read_routes(service, "/messages")
}

/// Create Axum router with only the write (mutation) endpoints for Message.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_message_write_routes(service: Arc<MessageService>) -> Router {
    BackboneCrudHandler::<
        MessageService,
        Message,
        CreateMessageDto,
        UpdateMessageDto,
        MessageResponseDto,
    >::write_routes(service, "/messages")
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_message_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<MessageService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_message_routes(service).layer(middleware::from_fn(
        move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req
                    .headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| {
                        raw.strip_prefix("Bearer ")
                            .or_else(|| raw.strip_prefix("bearer "))
                    })
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => (
                        axum::http::StatusCode::UNAUTHORIZED,
                        axum::Json(serde_json::json!({
                            "success": false,
                            "error": "unauthorized",
                            "message": "Authentication required"
                        })),
                    )
                        .into_response(),
                }
            }
        },
    ))
}